Split parser_v2 into constituent parts
