+++ cmake/objects.cmake
@@ -34,8 +34,15 @@ set(LIBDDWAF_SOURCE
     ${libddwaf_SOURCE_DIR}/src/parser/common.cpp
     ${libddwaf_SOURCE_DIR}/src/parser/parser.cpp
     ${libddwaf_SOURCE_DIR}/src/parser/parser_v1.cpp
-    ${libddwaf_SOURCE_DIR}/src/parser/parser_v2.cpp
     ${libddwaf_SOURCE_DIR}/src/parser/rule_data_parser.cpp
+    ${libddwaf_SOURCE_DIR}/src/parser/processor_parser.cpp
+    ${libddwaf_SOURCE_DIR}/src/parser/expression_parser.cpp
+    ${libddwaf_SOURCE_DIR}/src/parser/matcher_parser.cpp
+    ${libddwaf_SOURCE_DIR}/src/parser/transformer_parser.cpp
+    ${libddwaf_SOURCE_DIR}/src/parser/rule_parser.cpp
+    ${libddwaf_SOURCE_DIR}/src/parser/rule_override_parser.cpp
+    ${libddwaf_SOURCE_DIR}/src/parser/scanner_parser.cpp
+    ${libddwaf_SOURCE_DIR}/src/parser/exclusion_parser.cpp
     ${libddwaf_SOURCE_DIR}/src/processor/extract_schema.cpp
     ${libddwaf_SOURCE_DIR}/src/condition/lfi_detector.cpp
     ${libddwaf_SOURCE_DIR}/src/condition/sqli_detector.cpp
