+++ src/parser/rule_data_parser.hpp
@@ -11,8 +11,4 @@
 #include "parameter.hpp"
 #include "parser/common.hpp"
 
-namespace ddwaf::parser {
-
-template <typename T> T parse_rule_data(std::string_view type, parameter &input);
-
-}
+namespace ddwaf::parser {}
